feat: add postgreSQL integration

+++ tests/setup.js
@@ -5,9 +5,12 @@ import { createSchema } from "./packages/db/index.js";
 import { rootPath } from "./packages/utils/index.js";
 
 const setup = async () => {
-  const databasePath = path.join(rootPath, ".db-test");
-  if (fs.existsSync(databasePath)) {
-    fs.rmSync(databasePath, { recursive: true, force: true });
+  if (process.env.DATABASE_CLIENT === "sqlite") {
+    const dir = path.dirname(process.env.DATABASE_PATH);
+    const databasePath = path.join(rootPath, dir);
+    if (fs.existsSync(databasePath)) {
+      fs.rmSync(databasePath, { recursive: true, force: true });
+    }
   }
 
   await createSchema();
